+++ src/readers.rs
@@ -210,7 +210,7 @@ impl crate::FileFormat {
             } else if buffer.contains("<mxfile") {
                 return Ok(Self::Drawio);
             } else if buffer.contains("<x3d") {
-                return Ok(Self::Extensible3dGraphics);
+                return Ok(Self::Extensible3d);
             } else if buffer.contains("<xsl") {
                 return Ok(Self::ExtensibleStylesheetLanguageTransformations);
             } else if buffer.contains("<gml") {